refactor mappers
